// src/features/landing/types.d.ts
export interface Stat {
  value: string;
  label: string;
}

export interface Feature {
  icon: string;
  title: string;
  description: string;
}

export interface Service {
  icon: string;
  title: string;
  description: string;
  color: string;
}

export interface NavLink {
  label: string;
  href: string;
}